+++ CHANGELOG.md
@@ -2,9 +2,13 @@
 
 ## Unreleased
 
+### 🛑 Breaking changes 🛑
+
+- (Splunk) `confmap`: Do not expand special shell variable such as `$*` in configuration files. ([#5206](https://github.com/signalfx/splunk-otel-collector/pull/5206))
+
 ### 🚀 New components 🚀
 
-- (Splunk) Add `ngnix` receiver ([5229](https://github.com/signalfx/splunk-otel-collector/pull/5229))
+- (Splunk) Add `nginx` receiver ([5229](https://github.com/signalfx/splunk-otel-collector/pull/5229))
 
 ## v0.106.1